Clean compile, basic DHCPv6 functionality is there.
TODO
     hostname handling.
     update DHCP6 configs from dns
     parse domain=<domain>,<IPv6 range>
     pretty-print counted string options.
     DECLINE messages
     lease-script fro DHCPv6

+++ src/netlink.c
@@ -282,7 +282,9 @@ int iface_enumerate(int family, void *parm, int (*callback)())
 		rta = RTA_NEXT(rta, len1);
 	      }
 
-	    if (mac && !((*callback)(link->ifi_type, link->ifi_flags, mac, maclen, parm)))
+	    if (mac && !((*callback)((unsigned int)link->ifi_type, 
+				     (unsigned int)link->ifi_flags,
+				     mac, maclen, parm)))
 	      return 0;
 	  }
 #endif
